% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAFitAddFakeBulge.R
\name{GRAFitAddFakeBulge}
\alias{GRAFitAddFakeBulge}
\title{GRAFit: Add a Fake Bulge.}
\usage{
GRAFitAddFakeBulge(model = NULL, zeropoint = 34)
}
\arguments{
\item{model}{A list of the optimized model.}

\item{zeropoint}{Numeric scalar; the magnitude zero point.}
}
\value{
A model with a fake bulge model.
}
\description{
This low-level function suits the plotting process of the 1D profile when we have a single model. This actually added a very faint bulge to the model (fake bulge).
}
\examples{
-
}
\seealso{
\code{\link[GRAFit]{GRAFitDynamo_v2}}
}
\author{
Hosein Hashemizadeh
}
