% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAFitAddhdr.R
\name{GRAFitAddhdr}
\alias{GRAFitAddhdr}
\title{GRAFit: Add Header to the Output Catalogue.}
\usage{
GRAFitAddhdr(out_catalog)
}
\arguments{
\item{out_catalog}{The direction to the catalogue for which the header needs to be appended.}
}
\value{
Output catalogue.
}
\description{
This function adds a header to the output catalogue.
}
\examples{
-
}
\seealso{
\code{\link[GRAFit]{GRAFitfitClass}}, \code{\link[ProFit]{profitEllipsePlot}}
}
\author{
Hosein Hashemizadeh
}
