% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAFitAperturePlot.R
\name{GRAFitAperturePlot}
\alias{GRAFitAperturePlot}
\title{GRAFit: Overplot the Ellipsoidal Aperture.}
\usage{
GRAFitAperturePlot(x = x, y = y, a = a, b = b, angle = angle, border = "red")
}
\arguments{
\item{x, y}{The location of the aperture to be plotted.}

\item{a, b}{minor and major axes of the apperture.}

\item{angle}{Position angle}

\item{border}{The colour of the ellipse.}
}
\value{
Overplots the ellipsoidal apertur on the image.
}
\description{
This function overplots the ellipsoidal apertures on the image.
}
\examples{
-
}
\seealso{
\code{\link[GRAFit]{GRAFitEllipsePlot}}
}
\author{
Hosein Hashemizadeh
}
