% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAFitcutout.R
\name{GRAFitcutout}
\alias{GRAFitcutout}
\title{GRAFit: Image CutoutWCS}
\usage{
GRAFitcutout(
  image,
  loc = dim(image)/2,
  box = c(100, 100),
  shiftloc = FALSE,
  paddim = TRUE,
  padVal = 0,
  plot = FALSE,
  ...
)
}
\arguments{
\item{image}{Numeric matrix; required, the image we want to decorate. If image is a list as created by readFITS, read.fits of magcutoutWCS then the image part of these lists is parsed to image and the correct header part is parsed to header.}

\item{loc}{Numeric vector; the [x,y] (magcutout) or [x,y] / [RA,Dec] (magcutoutWCS) location where we want to cutout the image. For magcutoutWCS the unit type can be specified with the loc.type option. Either it is WCS in degrees [RA,Dec] (coord) or pixel [x,y] of the image (image).}

\item{box}{Numeric vector; the dimensions of the box to cut out from image centred on loc. For magcutout the box unit is always pixels. For magcutoutWCS the unit type can be specified with the loc.type option. Either it is pixels or asec (see loc.type).}

\item{shiftloc}{Logical; should the cutout centre shift from loc away from the image edge if the desired box size extends beyond the edge of the image?}

\item{paddim}{Logical; should the cutout be padded with image data until it meets the desired box size (if shiftloc is TRUE) or padded with NAs for data outside the image boundary otherwise?}

\item{padVal}{the valuse to be replaced if a cutout needed to be padded.}

\item{plot}{Logical; should a plot be generated?}

\item{...}{Dots are parsed to either \code{\link[magimage]{magimage}} (GRAFitcutout) or \code{\link[magimage]{magimageWCS}} (GRAFitcutoutWCS)}

\item{header}{Full FITS header in table or vector format. Legal table format headers are provided by the read.fitshdr function or the hdr list output of read.fits in the astro package; the hdr output of readFITS in the FITSio package or the header output of magcutoutWCS. If a header is provided then key words will be taken from here as a priority. Missing header keywords are printed out and other header option arguments are used in these cases.}

\item{CRVAL1}{FITS header CRVAL1 for the CTYPE1 projection system. This is the RA in degrees at the location of CRPIX1.}

\item{CRVAL2}{FITS header CRVAL2 for the CTYPE2 projection system. This is the Dec in degrees at the location of CRPIX2.}

\item{CRPIX1}{FITS header CRPIX1 for the CTYPE1 projection system. This is the x pixel value at the location of CRVAL1.}

\item{CRPIX2}{FITS header CRPIX2 for the CTYPE2 projection system. This is the y pixel value at the location of CRVAL2.}

\item{CD1_1}{FITS header CD1_1 for the CTYPE1 projection system. Change in CTYPE1 in degrees along x-Axis.}

\item{CD1_2}{FITS header CD1_2 for the CTYPE1 projection system. Change in CTYPE1 in degrees along y-Axis.}

\item{CD2_1}{FITS header CD2_1 for the CTYPE2 projection system. Change in CTYPE2 in degrees along x-Axis.}

\item{CD2_2}{FITS header CD2_2 for the CTYPE2 projection system. Change in CTYPE2 in degrees along y-Axis.}

\item{CTYPE1}{The RA projection system type. Either 'RA–TAN' for Tan Gnomonic (default), or 'RA–SIN' for Sine Orthographic. 'RA–NCP' is approximated by Sine Orthographic with a warning. Over-ridden by the FITS header.}

\item{CTYPE2}{The DEC projection system type. Either 'DEC–TAN' for Tan Gnomonic (default), or 'DEC–SIN' for Sine Orthographic. 'DEC–NCP' is approximated by Sine Orthographic with a warning. Over-ridden by the FITS header.}

\item{coord.type}{The units of loc for magcutoutWCS. Allowed options are 'deg' for degress and 'sex' for sexigesimal (i.e. HMS for RA and DMS for Deg).}

\item{sep}{When coord.type='sex', sep defines the type of separator used for the HMS and DMS strings (i.e. H:M:S and D:M:S would be sep=':', which is the default). See hms2deg and dms2deg for more details.}

\item{loc.type}{Character vector; specifies what type of location is being provided. The first element specifies the coordinate type for loc and the second element is the coordinate type for box. Either it is WCS in degrees [RA,Dec] / asec ('coord') or pixel [x,y] of the image ('image'). If only one element is provided then the same coordinate type is used for both loc and box.}
}
\value{
A file containing the coordination of each frame's center.
}
\description{
This function cuts a subset of an image with associated WCS system. This is originally developed as part of \code{\link[magicaxis]{magicaxis}} package. It has been modified to handle different paddings with arbitrary values.
}
\examples{
-
}
\seealso{
\code{\link[GRAFit]{GRAFitcutout}}
}
\author{
Aaron Robotham; Modified by: Hosein Hashemizadeh
}
