% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAFitDynamo_v2.R
\name{GRAFitDynamo_v2}
\alias{GRAFitDynamo_v2}
\title{GRAFit: Dynamic Cutout}
\usage{
GRAFitDynamo_v2(
  im = NULL,
  loc = c(ra_deg, dec_deg),
  init_xrad = NULL,
  init_yrad = NULL,
  R90 = NULL,
  finalCut = 5,
  finalBoxCar = 3,
  GRAFitlib = GRAFitlib,
  header = NULL,
  pix_scale = NULL,
  verbose = TRUE,
  magzero = NULL,
  tolerance = 7,
  smooth = TRUE,
  sigma = 7,
  sky = NULL,
  pixcut = 3,
  skycut = 1.1,
  boundstats = TRUE,
  rotstats = TRUE,
  ImPlot = TRUE,
  output_dir = NULL,
  ...
)
}
\arguments{
\item{im}{Image; a 2D image matrix.}

\item{loc}{Objects location; \code{c(RA,DEC)}}

\item{init_xrad}{Initial cutout size in x-direction in units of asec.}

\item{R90}{The radius containing 90\% of the total flux. In units of asec.}

\item{finalCut}{Size of the final cutout as finalCut*R90. Default = 5, i.e., 5*R90.}

\item{finalBoxCar}{Integer vector; the dimensions of the box car filter to estimate the sky with. Default = 3, i.e. 3*R90. For convenience, if length 1 then both dimensions of box used internally are assumed to equal the specified box. i.e. 200 would be interpreted as c(200,200). Dependent default arguments (grid, boxadd and skypixmin) are updated sensibly.}

\item{header}{Full FITS header in table or vector format.}

\item{pix_scale}{Pixel scale in units of arcsecond/pixel.}

\item{verbose}{verbose.}

\item{magzero}{Numeric scalar; the magnitude zero point.}

\item{tolerance}{Numeric scalar; the minimum height of the object in the units of skyRMS between its highest point (seed) and the point where it contacts another object (checked for every contact pixel). If the height is smaller than the tolerance, the object will be combined with one of its neighbours, which is the highest. The range 1-5 offers decent results usually. Passed to \code{\link[ProFound]{profoundMakeSegim}}.}

\item{smooth}{Logical; should smoothing be done on the target image? Passed to \code{\link[ProFound]{profoundMakeSegim}}.}

\item{sigma}{Numeric scalar; standard deviation of the blur used when smooth=TRUE. Passed to \code{\link[ProFound]{profoundMakeSegim}}.}

\item{sky}{User provided estimate of the absolute sky level. If this is not provided then it will be computed internally using \code{\link[ProFound]{profoundMakeSkyGrid}}. Can be a scalar or a matrix matching the dimensions of image (allows values to vary per pixel). This will be subtracted off the image internally, so only provide this if the sky does need to be subtracted!}

\item{pixcut}{Integer scalar; the number of pixels required to identify an object.}

\item{skycut}{Numeric scalar; the lowest threshold to make on the image in units of the skyRMS.}

\item{boundstats}{Logical; if TRUE then various pixel boundary statistics are computed (\code{Nedge}, \code{Nsky}, \code{Nobject}, \code{Nborder}, \code{edge_frac}, \code{edge_excess} and \code{FlagBorder}). If FALSE these return NA instead (saving computation time).}

\item{rotstats}{Logical; if TRUE then the \code{asymm}, \code{flux_reflect} and \code{mag_reflect} are computed, else they are set to NA. This is because they are very expensive to compute compared to other photometric properties.}

\item{ImPlot}{Logical; should the image be plotted?}

\item{output_dir}{The directory to save outputs.}
}
\value{
Plots as well we matrices of the image, segmentation, sky etc.
}
\description{
This high level function utilizes the effective radius (Re) of the object to dynamically make a cutout around it.
}
\examples{
-
}
\seealso{
\code{\link[GRAFit]{GRAFitLA}}, \code{\link[LaplacesDemon]{LaplacesDemon}}
}
\author{
Hosein Hashemizadeh
}
