% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAFitfitClass.R
\name{GRAFitfitClass}
\alias{GRAFitfitClass}
\title{GRAFit: Fit Class.}
\usage{
GRAFitfitClass(SB = NULL, SBlim = 26)
}
\arguments{
\item{SB}{A vector of surface brightness values.}

\item{SBlim}{5 sigma surface brightness limit of the data typically in units of mag/asec^2. Default = 26.}
}
\value{
An integer between 0-6. 0: normal fit, 1: bulge is likely fitting the outer halo or sky, 2: bulge is likely fitting truncated disk/outer region, 3: single component ellipticals or bulge trying to fit outer halo, 4: likely a clumpy disk with perturbations or strong feature of irregular galaxies, 5: likely a disk dominated (B/T < 0.5) or pure disk (B/T = 0) system. 6: likely a bulge dominated or pure bulge (B/T = 1) system. Likely an elliptical single sersic.
}
\description{
This function returns an integer indicating the fit class according to how many times the 1D profiles of the data and model cross.
}
\examples{
-
}
\seealso{
\code{\link[GRAFit]{GRAFitEllipsePlot}}
}
\author{
Hosein Hashemizadeh
}
