% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAFitFrameCenFinderRawFrame.R
\name{GRAFitFrameCenFinderRawFrame}
\alias{GRAFitFrameCenFinderRawFrame}
\title{GRAFit: Find the Center of Raw Image Frame.}
\usage{
GRAFitFrameCenFinderRawFrame(data_dir = NULL)
}
\arguments{
\item{data_dir}{The directory where the imaging data (frames) is stored.}
}
\value{
A file containing the coordination of each frame's center.
}
\description{
This function calculates the center of each image frame and saves coordinates into a file. This is useful for locating each galaxy between large number of frames.
}
\examples{
-
}
\seealso{
\code{\link[GRAFit]{GRAFitFrameCenFinder}}
}
\author{
Hosein Hashemizadeh
}
