% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAFitFrameFinder_v2.R
\name{GRAFitFrameFinder_v2}
\alias{GRAFitFrameFinder_v2}
\title{GRAFit: Find the frame within which the galaxy is located}
\usage{
GRAFitFrameFinder_v2(data_dir = NULL, target_loc = NULL, FrameType = "driz")
}
\arguments{
\item{data_dir}{The directory where the imaging data (frames) is stored.}

\item{target_loc}{target location; \code{c(RA,DEC)}.}

\item{FrameType}{\code{"driz"} for drizzled ACS/HST image or \code{"raw"} for raw ACS/HST image.}
}
\value{
A string specifying the directory/frame.name.
}
\description{
This function recieves the location (\code{RA,DEC}) of an object and returns the frame within which the galaxy is located. The frame can then be used for cutouts.
}
\examples{
-
}
\seealso{
\code{\link[GRAFit]{GRAFitFrameCenFinder}}, \code{\link[GRAFit]{GRAFitFrameCenFinderRawFrame}}
}
\author{
Hosein Hashemizadeh
}
