% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAFitInitModelSetupConst.R
\name{GRAFitInitModelSetupConst}
\alias{GRAFitInitModelSetupConst}
\title{GRAFit: Setup the Data/Model}
\usage{
GRAFitInitModelSetupConst(
  output_dir = NULL,
  GRAFitlib = GRAFitlib,
  main_src = NULL,
  nComp = 2,
  ExpDisk = FALSE,
  FreeBulge = FALSE,
  BulgeFreeness = 11,
  BT_flux_ratio = 0.2,
  ZP = NULL,
  image = image,
  mask = mask,
  sigma = sigma,
  pix_scale = pix_scale,
  segim = segim,
  psf = psf,
  SBlim = SBlim,
  like.func = "t",
  DoPriors = TRUE,
  DoConstraits = TRUE
)
}
\arguments{
\item{output_dir}{The directory to save outputs including the plots of the final model and standard \code{ProFit} plots.}

\item{main_src}{A list containing the specification of the main source in the cutout. This should be generated by the \code{ProFound}.}

\item{nComp}{Number of components to be used in fitting. Default = 2 for a bulge+disk model. Alternatively could be 1 for a single Sersic model.}

\item{ExpDisk}{A logical value. Should the disk be fitted with a pure exponential profile (\code{Sersic index = 1}). Default = \code{FALSE}; for a free Sersic profile.}

\item{FreeBulge}{Logical; Should the bulge location (\code{x \& y}) also be fitted by GRAFit. Default = \code{FALSE}, i.e. the bulge position will be fixed to the disk position.}

\item{BulgeFreeness}{In unit pixels. How many pixels the bulge position is allowed to be free from the centre of disk. Default = 11.}

\item{BT_flux_ratio}{The Bulge/Total flux ratio to estimate the initial bulge flux/magnitude from the total magnitude.}

\item{ZP}{Numeric scalar; the magnitude zero point.}

\item{image}{Image matrix; required, the galaxy image we want to fit a model to. The galaxy should be approximately central within this image.}

\item{mask}{Logical matrix; optional, non galaxy parts of the image to mask out, where 1 means mask out and 0 means use for analysis. If region is not provided then 0 values are used to define the common area to use for fitting. This matrix *must* be the same dimensions as image.}

\item{sigma}{Sigma matrix; optional, the measurement errors per pixel (expressed in terms of sigma). This matrix *must* be the same dimensions as image}

\item{pix_scale}{Pixel scale in units of arcsecond/pixel.}

\item{segim}{Segmentation matrix; optional, the full segmentation map of the image. If region is not provided then value of the central pixel is used to select the segmented pixels of the galaxy we want to fit. The log-likelihood is then computed using only these pixels. This matrix *must* be the same dimensions as image.}

\item{psf}{Matrix; optional. An empirical point spread function (PSF) image matrix that ProFit will use to convolve the image, as an alternative to defining an analytical PSF in modellist. This should have odd sizes in each dimension. If the dimension has an even size then the function will internally interpolate it onto an odd sized grid 1 element larger. \code{\link[ProFit]{profitSetupData}} forces negative values to equal 0. During any convolution \code{\link[ProFit]{profitConvolvePSF}} will force the sum of the pixels to equal 1 to ensure flux conservation during convolution of the model image.}

\item{SBlim}{5 sigma surface brightness limit of the data typically in units of mag/asec^2. Default = 26.}

\item{like.func}{likelihhod function to be parsed to the ProFit: "t": t-distribution, "norm": normal distribution. See ProFit package for more details.}

\item{DoPriors}{Logical; Should a prior distribution be applied to the fitting process. Default = \code{TRUE}.}

\item{DoConstraits}{Logical; Should constraints be applied to the fitting process. Default = \code{TRUE}.}
}
\value{
Plots of the initial model as well as the Data as a list.
}
\description{
This function sets up the profit.class data and the initial model required for optimization.
}
\examples{
-
}
\seealso{
\code{\link[GRAFit]{GRAFitDynamo_v2}}
}
\author{
Hosein Hashemizadeh
}
