% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAFitLA.R
\name{GRAFitLA}
\alias{GRAFitLA}
\title{GRAFit: perform Laplace Approximation}
\usage{
GRAFitLA(
  Model = profitLikeModel,
  Initial.Values = Data$init,
  Data = Data,
  iteration = 1000,
  Method = "LM",
  verbose = TRUE
)
}
\arguments{
\item{Model}{profitLikeModel}

\item{Initial.Values}{Initial values.}

\item{Data}{Data of class profit.data. The standard Data structure containing all inputs, images, modelsetup etc.}

\item{iteration}{An integer. The iterations for optimization. Default = 1000}

\item{Method}{A string that specifies the mothod for Laplace Approximation. Default = \code{'LM'} for Levenberg-Marquardt. See \code{\link[LaplacesDemon]{LaplaceApproximation}} for all options.}

\item{verbose}{verbose.}
}
\value{
A list of optimized fit containing both the output of \code{LaplaceApproximation} and \code{profitRemakeModellist}.
}
\description{
This low level function acts as a wrapper around the \code{\link[LaplacesDemon]{LaplaceApproximation}}.
}
\examples{
-
}
\seealso{
\code{\link[GRAFit]{GRAFitLD}}, \code{\link[LaplacesDemon]{LaplaceApproximation}}
}
\author{
Hosein Hashemizadeh
}
