% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAFitLD.R
\name{GRAFitLD}
\alias{GRAFitLD}
\title{GRAFit: perform MCMC using LaplacesDemon}
\usage{
GRAFitLD(
  output_dir = output_dir,
  GRAFitlib = GRAFitlib,
  Model = profitLikeModel,
  nComp = 2,
  Initial.Values = NULL,
  Data = Data,
  iteration = 10000,
  zeropoint = ZP,
  pixscale = 0.03,
  FWHM = 0.09,
  SBlim = 26,
  Algorithm = "CHARM",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{output_dir}{The directory to save outputs including the plots of the final model and standard \code{ProFit} plots.}

\item{Model}{profitLikeModel}

\item{nComp}{Number of components to be used in fitting. Default = 2 for a bulge+disk model. Alternatively could be 1 for a single Sersic model.}

\item{Initial.Values}{A vector of initial values equal in length to the number of parameters.}

\item{Data}{Data of class profit.data. The standard Data structure containing all inputs, images, modelsetup etc.}

\item{iteration}{The iterations for \code{MCMC} optimization if \code{optimMode = 'MCMC'}. Default = 10000}

\item{zeropoint}{Numeric scalar; the magnitude zero point.}

\item{pixscale}{Pixel scale in arcsecond/pixel.}

\item{FWHM}{The full width half max of the PSF in units of arc seconds. A vertical line is drawn at half this number (since we are plotting radius). The fits inside of the region is inherently hard because it is well within the PSF convolution kernel}

\item{SBlim}{5 sigma surface brightness limit of the data typically in units of mag/asec^2. Default = 26.}

\item{Algorithm}{The MCMC algorithm to be used for the optimization. Default = 'CHARM'. You probably don't need to change this as CHARM is shown to be very robust algorithm.}

\item{verbose}{verbose.}
}
\value{
Plots and optimized model.
}
\description{
This low level function acts as a wrapper around the \code{LalacesDemon} package to perform an MCMC sampling.
}
\examples{
-
}
\seealso{
\code{\link[GRAFit]{GRAFitLA}}, \code{\link[LaplacesDemon]{LaplacesDemon}}
}
\author{
Hosein Hashemizadeh
}
