% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAFitMainFinder.R
\name{GRAFitMainFinder}
\alias{GRAFitMainFinder}
\title{GRAFit: Find the Target Object in the Cutout.}
\usage{
GRAFitMainFinder(src_list = NULL, imDim = NULL, main_pin = FALSE)
}
\arguments{
\item{src_list}{The output of \code{\link[ProFound]{ProFound}} source detection.}

\item{imDim}{An n*m matrix. The dimension of the cutout.}

\item{main_pin}{Should the location of the main object on the cutout be shown by a symbol. This can only be \code{TRUE} if the image is open.}
}
\value{
A file containing the coordination of each frame's center.
}
\description{
This function finds the main target/galaxy a cutout.
}
\examples{
-
}
\seealso{
\code{\link[GRAFit]{GRAFitDynamo_v2}}
}
\author{
Hosein Hashemizadeh
}
