% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAFitMaster.R
\name{GRAFitMaster}
\alias{GRAFitMaster}
\title{GRAFit: Master code for connecting and calling other modules.}
\usage{
GRAFitMaster(
  wrk_dir = NULL,
  data_dir = NULL,
  GRAFitlib = NULL,
  PSF_dir = NULL,
  object_list = NULL,
  threadMode = c(0, 1),
  ncores = 1,
  logfile = "logfile.txt",
  nComp = 2,
  optimMode = "MCMC",
  LA_iteration = 1000,
  MCMC_iteration = 10000,
  ExpDisk = FALSE,
  FreeBulge = FALSE,
  BulgeFreeness = 11,
  Single_PSF = FALSE,
  PSF_diameter = 1,
  like.func = "t",
  catalog_name = "MasterCat.csv",
  keep_wrk_space = FALSE,
  verbose = TRUE,
  plot = FALSE,
  add_hdr = TRUE,
  DoPriors = TRUE,
  DoConstraits = TRUE
)
}
\arguments{
\item{wrk_dir}{Working directory, where the GRAFit outputs will be saved.}

\item{data_dir}{The directory where the imaging data (frames) is stored.}

\item{PSF_dir}{The directory where pre-generated PSFs are stored. If this is not provided the PSF can be put into the wrk_dir. Alternatively, GRAFit will generate PSF for each galaxy.}

\item{object_list}{A catalogue of objects you wish to fit, should include: CATAID, RA, DEC, etc.}

\item{threadMode}{specifies the thereading mode of the parallel mode of the GRAFit. 0: snow-like, suitable for laptops 1: using Rmpi, suitable for large clusters like supercomputers.}

\item{ncores}{Number of CPUs to be used in parallel mode. Default = 1}

\item{logfile}{Logical; Should a log file be generated. Default = \code{TRUE}}

\item{nComp}{Number of components to be used in fitting. Default = 2 for a bulge+disk model. Alternatively could be 1 for a single Sersic model.}

\item{optimMode}{The optimization mode. Default = 'MCMC'; using \code{LaplacesDemon} package. Alternatives for this argument is: 'LA' that uses \code{Laplace Approximation} and 'optim' that uses \code{optim} optimization.}

\item{LA_iteration}{The iterations for \code{Laplace Approximation} optimization if optimMode = 'LA'. Default = 1000}

\item{MCMC_iteration}{The iterations for \code{MCMC} optimization if optimMode = 'MCMC'. Default = 10000}

\item{ExpDisk}{Logical; Should the disk be fitted with a pure exponential profile (\code{Sersic index = 1}). Default = \code{FALSE}; for a free Sersic profile.}

\item{FreeBulge}{Logical; Should the bulge location (\code{x \& y}) also be fitted by GRAFit. Default = \code{FALSE}, i.e. the bulge position will be fixed to the disk position.}

\item{BulgeFreeness}{In unit pixels. How many pixels the bulge position is allowed to be free from the centre of disk. Default = 11.}

\item{Single_PSF}{If \code{PSF_dir = FALSE} should the PSF be generated as a single PSF on the position of the galaxy on the CCD. Default = \code{FALSE}, generate a stacked PSF on the location of galaxy on each of the raw ACS images.}

\item{PSF_diameter}{the diameter of the PSF in arcseconds. Default = 1"}

\item{like.func}{likelihhod function to be parsed to the ProFit: "t": t-distribution, "norm": normal distribution. See ProFit package for more details.}

\item{catalog_name}{The name of GRAFit's output catalogue.}

\item{keep_wrk_space}{Logical; Should the work space be saved. Default: \code{TRUE}.}

\item{verbose}{Logical; Verbose.}

\item{plot}{Logical; Should plots be generated or just the structural catalogue.}

\item{add_hdr}{Logical; Should a header be appended to the top of the catalogue. Default = \code{TRUE}. Usefull for some large runs on supercomputers large dataset is splitted into several chunks.}

\item{DoPriors}{Logical; Should a prior distribution be applied to the fitting process. Default = \code{TRUE}.}

\item{DoConstraits}{Logical; Should constraints be applied to the fitting process. Default = \code{TRUE}.}
}
\value{
GRAFit's output is a structural catalogue (catalog_name.csv) as well as a folder named with the object's ID containing plots and figures of the initial and final models.
}
\description{
This code is considered as the core code of the GRAFit that connects all other parts and calls modules where needed.
}
\examples{
GRAFitMaster(wrk_dir = '~/Desktop/wrk_dir/', data_dir = '~/Desktop/data_dir/', PSF_dir = '~/Desktop/PSF_dir/', threadMode = 0, ncores = 1, nComp= 2, optimMode = 'MCMC', object_list = 'object_list')
}
\seealso{
\code{\link[GRAFit]{GRAFitDynamo_v2}}
}
\author{
Hosein Hashemizadeh
}
