% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAFitmu0.R
\name{GRAFitMu0}
\alias{GRAFitMu0}
\title{GRAFit: calculates the central surface brightness}
\usage{
GRAFitMu0(n, mag, Re, PixScale = 1)
}
\arguments{
\item{n}{Sersic index}

\item{mag}{Magnitude}

\item{Re}{The effective radius either in pixel for which the PixScale should be provided or arcsecond.}

\item{PixScale}{Pixel scale in arcsecond/pixel. \code{PixScale = 1} (default) means that Re is in the units of arcsecond. Should be necessarily provided if the Re is in units of pixels.}
}
\value{
The central surface brightness of galaxy in units of mag/asec^2
}
\description{
This function calculates the central surface brightness of a galaxy from its Sersic index profile using the euation in Graham & Driver 2005.
}
\examples{
-
GRAFitMu0(n = 2, mag = 21, Re = 20, PixScale = 0.03)
}
\references{
Graham A. W., Driver S. P., 2005, PASA, 22, 118
}
\seealso{
\code{\link[GRAFit]{GRAFitEllipsePlot}}
}
\author{
Hosein Hashemizadeh
}
