% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAFitOptim.R
\name{GRAFitOptim}
\alias{GRAFitOptim}
\title{GRAFit: perform optim fit}
\usage{
GRAFitOptim(
  output_dir = output_dir,
  Model = profitLikeModel,
  nComp = 2,
  Initial.Values = NULL,
  Data = Data,
  Algorithm = "BFGS",
  zeropoint = ZP,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{output_dir}{The directory to save outputs including the plots of the final model and standard \code{ProFit} plots.}

\item{Model}{profitLikeModel}

\item{nComp}{Number of components to be used in fitting. Default = 2 for a bulge+disk model. Alternatively could be 1 for a single Sersic model.}

\item{Initial.Values}{Initial values.}

\item{Data}{Data of class profit.data. The standard Data structure containing all inputs, images, modelsetup etc.}

\item{Algorithm}{A string that specifies the algorithm for optimization. Default = \code{'BFGS'} a quasi-Newton method introduced by Broyden, Fletcher, Goldfarb and Shanno. Other options \code{"Nelder-Mead"}, \code{"CG"}, \code{"L-BFGS-B"}, \code{"SANN"}, \code{"Brent"}. See \code{\link[stats]{optim}} for more details.}

\item{zeropoint}{Numeric scalar; the magnitude zero point.}

\item{verbose}{verbose.}

\item{...}{extra parameters to be parsed to \code{\link[GRAFit]{GRAFitEllipsePlot}}.}
}
\value{
A list of optimized fit containing both the output of \code{LaplaceApproximation} and \code{profitRemakeModellist}.
}
\description{
This low level function acts as a wrapper around the \code{\link[stats]{optim}}. It can handle a few optimization algorithms including: \code{"Nelder-Mead"}, \code{"CG"}, \code{"BFGS"}, \code{"L-BFGS-B"}, \code{"SANN"}, \code{"Brent"}.
}
\examples{
-
}
\seealso{
\code{\link[GRAFit]{GRAFitLD}}, \code{\link[GRAFit]{GRAFitLA}}, \code{\link[GRAFit]{GRAFitcutoutWCS}}
}
\author{
Hosein Hashemizadeh
}
