% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAFitPSFgenerator.R
\name{GRAFitPSFgenerator}
\alias{GRAFitPSFgenerator}
\title{GRAFit: PSF Generator}
\usage{
GRAFitPSFgenerator(
  wrk_dir = wrk_dir,
  output_dir = output_dir,
  HST_Focus_val_file = NULL,
  PSF_name = "psf",
  target_loc,
  loc_Unit = c("xy", "deg"),
  CCDCHIP,
  filter = "f814w",
  EXPSTART,
  jitter = 3,
  header = NULL,
  ebmv = 0,
  SUB = 1,
  PSF_diameter = 1,
  down_sample_factor,
  verbose = TRUE
)
}
\arguments{
\item{wrk_dir}{Working directory, where the GRAFit outputs will be saved.}

\item{output_dir}{The directory to save the PSF.}

\item{HST_Focus_val_file}{The directory to the file where the HST focus values are stored.}

\item{PSF_name}{The name of the output PSF.}

\item{target_loc}{Location; \code{c(RA,DEC)}.}

\item{loc_Unit}{target_loc unit; \code{"xy"}: CCD coordination or \code{"deg"}: celestial coordination.}

\item{CCDCHIP}{Which ACS's CCD chip the object is located in. Allowed values: 1 and 2.}

\item{filter}{What filter is your imaging data. E.g., \code{"f814w"}. If missing it will be read from the \code{header}.}

\item{EXPSTART}{Time of the start of exposure.}

\item{jitter}{Telescope's jitter. Default = 3.}

\item{header}{Header of images.}

\item{ebmv}{B-V extinction.}

\item{SUB}{Subsampling ratio to be used by the Tiny Tim.}

\item{PSF_diameter}{The diameter of the PSF in arcseconds. Default = 1"}

\item{down_sample_factor}{A factor by which the user wish the Tiny Tim PSF to be resampled.}

\item{verbose}{Logical; Verbose.}
}
\value{
The image matrix of the PSF.
}
\description{
This high-level function generates the HST PSF.
}
\examples{
-
}
\seealso{
\code{\link[GRAFit]{TinyTimACSR}}
}
\author{
Hosein Hashemizadeh
}
