% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAFitPSFMakeStack.R
\name{GRAFitPSFMakeStack}
\alias{GRAFitPSFMakeStack}
\title{GRAFit: Stack Four HST/ACS PSFs}
\usage{
GRAFitPSFMakeStack(psf1, psf2, psf3, psf4)
}
\arguments{
\item{psf1, psf2, psf3, psf4}{PSFs}
}
\value{
A stacked PSF.
}
\description{
This function stacks four HST/ACS PSFs.
}
\examples{
-
}
\seealso{
\code{\link[GRAFit]{GRAFitPSFgenerator}}, \code{\link[GRAFit]{GRAFitTinyTimACS}}
}
\author{
Hosein Hashemizadeh
}
