% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAFitri.R
\name{GRAFitri}
\alias{GRAFitri}
\title{GRAFit: MCMC Chain Corner Plot.}
\usage{
GRAFitri(chains, samples, samptype = "end", grid = FALSE, tick = FALSE)
}
\arguments{
\item{chains}{A matrix or data.frame of the posterior chains, arranged so that the columns are the parameters and rows are the individual chain samples. The column names are inherited as the parameter names from the input to chains.}

\item{samples}{Specify the number of sub-samples desired. To speed up plotting it is often a good idea not to plot all chain samples (the reduced set is plotted as the top-left points and used to generate the bottom-right contours). The default plot 1000 samples.}

\item{samptype}{Specifies whether to take all of the samples from the end of the supplied chains ('end', the default since samples are usually better towards the end of a set of psoterior chain samples), randomly selected ('ran', should only be used if you are confident the posterior chains supplied are true stationary samples) or evenly selected ('thin', which behaves much like thinning except we specify the target number of outputs, not the fraction of samples kept).}

\item{grid}{Logical; should a background grid be added to the sub-panels? See \code{\link[magicaxis]{magaxis}} for details.}

\item{tick}{Logical; should tick marks be added to every sub-panel?}
}
\value{
A corner plot (aka triangle plot).
}
\description{
A very high level (minimal options) MCMC chain triangle (AKA corner) plot function. The default is deliberately spartan in terms of options, but the result should be a clear set of covariance plots that should give quick insight into the stationary sampling quality of a set of MCMC posterior chains. This function is a modification (mostly cosmetic) of the \code{\link[magtri]{magtri}} from \code{\link[magicaxis]{magicaxis}} package.
}
\examples{
Sigma = matrix(c(10,3,-5,3,12,8,-5,8,20),3,3)
chains = MASS::mvrnorm(n = 1000, mu = 1:3, Sigma = Sigma)
GRAFitri(chains,tick = TRUE)
}
\seealso{
\code{\link[GRAFit]{GRAFitLD}}
}
\author{
Aaron Robotham; Modified for GRAFit by: Hosein Hashemizadeh
}
