% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAFitSBprofile.R
\name{GRAFitSBprofile}
\alias{GRAFitSBprofile}
\title{GRAFit: Pixel-by-pixel Surface Brightness Profile Plot.}
\usage{
GRAFitSBprofile(
  image = image,
  main_source = main_src,
  model = NULL,
  segim = segim,
  comp = c("bd", "b", "d"),
  centerPos = c(optim_xcen1, optim_ycen1),
  pixel_scale = 0.03,
  zeropoint = NULL,
  plot = TRUE,
  modelPlot = TRUE,
  title = NULL,
  col = NULL,
  legend = TRUE,
  legend_lab = c("Data", "Model+Noise", "Bulge", "Disk"),
  legend_col = c("grey", "green", "red", "blue")
)
}
\arguments{
\item{image}{Image matrix; required, the galaxy image we want to fit a model to. The galaxy should be approximately central within this image.}

\item{main_source}{A list containing the specification of the main source in the cutout. This should be generated by the \code{ProFound}.}

\item{model}{The matrix of the model.}

\item{segim}{Segmentation matrix; optional, the full segmentation map of the image. If region is not provided then value of the central pixel is used to select the segmented pixels of the galaxy we want to fit. The log-likelihood is then computed using only these pixels. This matrix *must* be the same dimensions as image.}

\item{comp}{The components to be plotted; \code{bd}: bulge+disk, \code{b}: bulge only, \code{bd}: disk only.}

\item{centerPos}{The position of the center; i.e., c(x,y)}

\item{zeropoint}{Numeric scalar; the magnitude zero point.}

\item{plot}{Logical; should the plot be generated? Otherwise only the values (SBs) will be returned.}

\item{modelPlot}{Logical; should the model profile be plotted?}

\item{title}{String; plot's title.}

\item{col}{A vector of colours.}

\item{legend}{Should a legend be annotated to the plot?}

\item{legend_lab}{A vector of legends.}

\item{legend_col}{A vector of legend's colours.}

\item{pix_scale}{Pixel scale in units of arcsecond/pixel.}
}
\value{
The pixel-by-pixel SB plot.
}
\description{
This high-level function calculates/plots the pixel-by-pixel surface brightness (SB).
}
\examples{
-
}
\seealso{
\code{\link[GRAFit]{GRAFitEllipsePlot}}
}
\author{
Hosein Hashemizadeh
}
