% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAFitTinyTimACS.R
\name{GRAFitTinyTimACS}
\alias{GRAFitTinyTimACS}
\title{GRAFit: Setup Tiny Tim Commands for PSF Generation}
\usage{
GRAFitTinyTimACS(
  wrk_dir = NULL,
  output_dir = NULL,
  name,
  CCDCHIP,
  x,
  y,
  filter = "f814w",
  focus,
  ebmv = 0,
  jitter = 3,
  PSF_diameter = 1,
  SUB = 1
)
}
\arguments{
\item{wrk_dir}{Working directory, where the GRAFit outputs will be saved.}

\item{output_dir}{The directory to save the PSF.}

\item{name}{The name of output file}

\item{CCDCHIP}{Which ACS's CCD chip the object is located in. Allowed values: 1 and 2.}

\item{x, y}{target_loc unit; CCD coordination.}

\item{filter}{What filter is your imaging data. E.g., \code{"f814w"}. If missing it will be read from the \code{header}.}

\item{focus}{HST focus values.}

\item{ebmv}{B-V extinction.}

\item{jitter}{Telescope's jitter. Default = 3.}

\item{PSF_diameter}{The diameter of the PSF in arcseconds. Default = 1".}

\item{SUB}{Subsampling ratio to be used by the Tiny Tim.}
}
\value{
A bash file with Tiny Tim commands.
}
\description{
Low-level function to setup the requiered Tiny Tim commands for HST PSF generation.
}
\examples{
-
}
\seealso{
\code{\link[GRAFit]{GRAFitPSFgenerator}}
}
\author{
Hosein Hashemizadeh
}
